+++ api/index.js
@@ -1,5 +1,5 @@
 import express from "express"
-import userRoutes from "./routes/employees.js"
+import employeeRoutes from "./routes/employees.js"
 import cors from "cors"
 
 const app = express()
@@ -7,6 +7,6 @@ const app = express()
 app.use(express.json())
 app.use(cors())
 
-app.use("/", userRoutes)
+app.use("/", employeeRoutes)
 
 app.listen(8800)
